+++ core-api/src/main/java/com/optimizely/ab/Optimizely.java
@@ -21,8 +21,8 @@ import com.optimizely.ab.bucketing.DecisionService;
 import com.optimizely.ab.bucketing.FeatureDecision;
 import com.optimizely.ab.bucketing.UserProfileService;
 import com.optimizely.ab.config.*;
-import com.optimizely.ab.config.optimizely.OptimizelyConfig;
-import com.optimizely.ab.config.optimizely.OptimizelyConfigService;
+import com.optimizely.ab.optimizelyconfig.OptimizelyConfig;
+import com.optimizely.ab.optimizelyconfig.OptimizelyConfigService;
 import com.optimizely.ab.config.parser.ConfigParseException;
 import com.optimizely.ab.error.ErrorHandler;
 import com.optimizely.ab.error.NoOpErrorHandler;
