+++ core-api/src/main/java/com/optimizely/ab/optimizelyconfig/OptimizelyConfigService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and      *
  * limitations under the License.                                           *
  ***************************************************************************/
-package com.optimizely.ab.config.optimizely;
+package com.optimizely.ab.optimizelyconfig;
 
 import com.optimizely.ab.config.*;
 import java.util.*;
